Add support for DISTINCT Keyword
Motivation
----------
Adds support for the DISTINCT keyword so that the duplicates can be
omitted from a result set.

Modifications
-------------
Made changes to the N1QLQueryModelVistor to recognize the expression
contains a DistinctResultOperator clause. Modifed the QueryPartsAggregator
to add the DISTINCT keyword to the rendered query.

Result
------
The provider now supports the Linq Distinct() method.