Add support for EXPLAIN keyword
Motivation
----------
The N1QL language supports the EXPLAIN keyword, which gives the query plan
for the N1QL query as JSON document.

Modifications
-------------
Added ExplainExpressionNode parser and extension method support for
IQueryable.

Result
------
Add query generator support for the N1QL EXPLAIN keyword. As it stands now, the
Linq provider will not know how to handle a result that doesn't match the
standard error response or target type of the Linq query. This will be
added in a subsquent commit.