Support automatic Where filters on entities based on their type
Motivation
------------
When working with LINQ, the expectation of most developers is that when querying using a particular model type only results of that model type are returned.  If results of other model types are returned from the bucket and mapped to properties of the wrong type, it can cause unexpected confusion and run-time errors.

Modifications
---------------
Provide a static EntityFilterManager class which manages filters to be applied to all LINQ-based queries, based on the model type being queried.  This can be managed at runtime using code, or more simply by applying attributes to the model classes.

Both of the Queryable<T>() implementations now automatically call the EntityFilterManager to apply where clauses to every query, when applicable.

Additionally, the EntityTypeFilterAttribute allows the developer to easily apply the most common filter, which is on the "type" attribute of the JSON object.

Results
--------
The developer can more seamlessly use Linq2Couchbase queries to access their data, using declarative attributes or more advanced runtime filtering.  The usage of LINQ is more intuitive to the developer and less likely to generate runtime errors.