Basic proof of concept unit tests for entity filters
Provides a test of entity filtering against the `beer-sample` bucket using both a declarative attribute and as well as a runtime filter.  Test ensures that unexpected entities are not returned by the query.

Note: Due to current limitations, the .All() method is not working with Linq2Couchbase.  It was therefore necessary to apply .AsEnumerable() first for these tests to function.