Improve generation method of FROM clause
Motivation
----------
Will allow for implementation of subqueries by removing the requirement
for a bucket name as the from clause in the N1QLQueryModelVisitor.

Modifications
-------------
VisitMainFromClause now recognizes that the FromExpression contains an
IBucketQueryable, and extracts the bucket name from there.  The bucket
name is no longer stored as a property of N1QLQueryModelVisitor.

FormatBucketName is replaced by EscapeIdentifier, and is used for
formatting both bucket names and other identifiers such as member names.

Results
-------
Unit tests still pass the same requirements as before the modifications.
It will now be easier to implement subqueries in the future, such as the
use of .Any() and .All() operations.  Also supports the use of generated
item names such as <generated>_1 by escaping them with ticks, i.e.
`<generated>_1`.