Implement .Any() subquery for nested arrays
Motivation
------------
Provide support for the N1QL ANY x IN y SATISFIES z END collection
operator, allowing the query to test predicates on collections nested
inside document objects.

Modifications
---------------
The QueryPartsAggregator now implements a QueryType property, which is set
to N1QlQueryPart.Any when the query being generated is an Any subquery.
To support generating these subqueries, the FromParts property is now a
collection of N1QlFromQueryPart objects rather than a collection of
strings.  This allows the from part to be formatted as "ANY <name> IN
<source>" when performing an Any subquery, or format in the traditional
"<source> as <name>" for the primary select query.

Results
--------
It is now possible to perform an .Any() or .Any(predicate) sub query on a
collection property of the main document.

Notes
-------
According to N1QL documentation at
http://docs.couchbase.com/4.0/n1ql/n1ql-language-reference/collectionops.html,
the SATISFIES clause should be optional.  However, in testing that does
not currently appear to be the case.  Therefore, "SATISFIES true" is being
used to test for the presence of items when performing .Any() without a
predicate.