Redesign META keyword to be invoked by a static method
Motivation
----------
The previous implementation of .Meta() in the LINQ chain could only be
invoked in the select statement.  Additionally, it was difficult to get
the metadata into the document entity.

Modifications
-------------
Removed the MetaResultOperator and .Meta() IQueryable extension.
Implemented a new N1Ql static class with N1Ql.Meta(document) as a static
method.  Implemented IDocumentMetadataProvider interface to support faking
metadata in unit tests.

Results
-------
Metadata can now be included directly in the "select" portion of a
Linq2Couchbase query.  You can also access properties of the metadata such
as the id in the query.  Finally, you can use the metadata in Where
predicates as well.

Example
-------
var query = from p in bucket.Queryable<Beer>
    select new {name = p.Name, id = N1Ql.Meta(p).Id}