Support for INNER JOIN, LEFT JOIN, INNER UNNEST, and OUTER UNNEST
Modifications
-------------
Handles join, group join, and additional from clauses in
N1QlQueryModelVisitor.  Throws exceptions for some known cases that are
unsupported.

Results
-------
Basic joins and unnests are now supported.  Cases where a Where clause is
applied to the inner sequence before the join operation are also handled
by moving the where clause to the main query.

Joins must always be made using the N1Ql.Key static method for the inner
key selector.  This is consistent with the N1QL requirement that joins
must use the ON KEYS statement.

Notes
-----
There are probably still edge cases that are not supported, some of which
are not throwing exceptions.

If you apply a Where clause to the inner sequence before a
DefaultIfEmpty() clause, the behavior between N1QL and LINQ is currently
inconsistent.  The generated N1QL query is filtering the Where clause
after the left join, while LINQ is filtering before the join.  Therefore,
it is possible that data on the left side that should be returned with a
null right side is dropped instead.