Fix bug where exception is thrown and error list is empty
Motivation
----------
If an exception is thrown and handled internally, we want the original
exception to be rethrown so its easier to track the issue.

Modifications
-------------
Updated the BucketQueryExecuter so that the original error is thrown.

Results
-------
If a timeout error occurs for example, the stacktrace and error will be
thrown, not a generic Exception.