Support for string concatentation N1QL operators
Motivation
----------
Support for string concatenation in queries is particularly important for
joins, since it can be used to build compound keys if the key is not
stored raw in the main document.

Modifications
-------------
Added ConcatMethodCallTranslator as well as string detection on the +
operator.  Both use the "||" N1QL concatenation operator.

Result
------
Supports string concatenation in LINQ queries using String.Concat()
overloads and the + operator.

Note
----
Based on work from the Relinq-SqlBackend example project at
https://github.com/re-motion/Relinq-SqlBackend.