Support for array definitions in query
Motivation
----------
Add support for the [] array definition N1QL operators

Modifications
-------------
Implemented handling for array ConstantExpression, as well as the
NewArrayInit expression for non-constant arrays.  Becasue the array
definition contains commas, had to change the SelectParts handling of
QueryPartsAggregator and N1QlQueryModelVisitor so that it no longer
attempts to split based on commas.  Now it works on the whole select part
string.

Results
-------
Queries can now define arrays of constants or arrays of query values in
the select part or the where part of the query.