Support for IS MISSING and IS VALUED operators
Motivation
----------
The current .WhereMissing() implementation is limited to the where clause,
and only does simple filtering logic.  It might be desired to do more
advanced operations with the IS MISSING and IS VALUED operators, such as
using them in CASE statements.

Modifications
-------------
Added IsMissing, IsNotMissing, IsValued, and IsNotValued static methods to
the N1Ql class.  Added method call translators for these methods.

Results
-------
It is now possible to generate IS MISSING, IS NOT MISSING, IS VALUED, and
IS NOT VALUED clauses in any expression in the query.  Using the
"propertyName" parameter string allows testing a property that doesn't
exist in the .Net object model.