Support for coalesing with ISMISSINGORNULL function
Motivation
----------
The C# ?? coalesce operator is a good match for the N1QL ISMISSINGORNULL
function.

Modifications
-------------
Coalesce expression are passed through the VisitCoaleseExpression method,
which recurses on the right side to build a comma-delimited list of
parameters for the ISMISSINGORNULL function.  Where any other expression
type is encountered, it is passed through the normal VisitExpression
method.