Support for bucket subqueries
Motivation
----------
Add the ability to query against buckets as a subquery.  This includes
testing for Any and All, as well as including the results in the
projection.

Modifications
-------------
Changed N1QlExtentNameProvider to return extent names already escaped.
This should improve performance, but also allowed the implementation of a
Prefix property.  This was necessary to automatically prefix extent name
references with another extent name when performing SATISFIES clauses.

Added SubqueryMethodCallTranslator, which basically ignores calls to
ToList(), ToArray(), and AsEnumerable().  Making one of these calls is
required when putting subquery results in the projection, so that the
return type is not IQueryable.

Made necessary changes to N1QlQueryModelVisitor and QueryPartsAggregator
to support bucket subqueries.  Also added isSubQuery as a constructor
parameter to N1QlQueryModelVisitor, used when creating a subquery model
visitor in N1QlExpressionTreeVisitor.

Results
-------
Subqueries against buckets are now supported in select and where clauses.

Notes
-----
A subquery in the select projection can really be an inefficient version
of a nest.  If the subquery is a simple query, without a select clause or
body clauses such as sorting or filtering, that is technically the
equivalent of a LEFT OUTER NEST.  It would be nice to recognize this
situation and automatically convert to the more efficient syntax in the
future.