Remove internal dependencies on Couchbase from project
Motivation
----------
Removed the dependencies on internal Types from the Couchbase client
project so that internalsVisibleTo no longer has to be used on the parent
project.

Modifications
-------------
Deleted QueryClientExtensions and related tests.

Results
-------
The parent project no longer has to use internalsVisibleTo the Couchbase
linq project; this makes it easier to package because we don't have to
share signing keys.