Improve N1QL identifier escaping performance
Modifications
-------------
Changed N1QlHelpers.EscapeIdentifier to using string.Concat instead of a
StringBuilder to wrap identifiers in tick marks.  When repeating the call
to EscapeIdentifier 10,000,000 times, it was approximately 50% slower to
use a StringBuilder for this function.  string.Concat, using strings and
not characters for the ticks, was the fastest alternative found.