Support for array property filtering and projection
Motivation
----------
Allow arrays within documents to be filtered or projected in the result
set.  This eliminates the need to perform this work on the client side,
and allows the filtered or projected array to be used in where predicates.

Modifications
-------------
Implemented handling in QueryPartsAggregator for the Array query type.
Throw exceptions in N1QlQueryModelVisitor if clauses are visited which
cannot be supported by an ARRAY clause in N1QL.

Results
-------
Array properties within documents can now be filtered and projected in any
part of the N1QL query.