Support for array sorting
Motivation
----------
Allow array properties or their filtered projections to be sorted.

Modifications
-------------
Removed exception when OrderByClause is reached for an array subquery.
Instead, only throw an exception if the OrderByClause is not a simple case
such as `.OrderBy(p => p)`.  Then wrap the array subquery in an ARRAY_SORT
function, which is the N1QL equivalent.  For descending sorts, wrap again
with an ARRAY_REVERSE function call.

Also, simplify subqueries which don't perform projections or filtering by
directly referencing the array property inside the wrapping functions.
The ARRAY clause is unnecessary in this case.

Results
-------
Applying OrderBy clauses to arrays inside documents now works, so long as
you are sorting by the array elements themselves.