LINQ-8: Add DbContext like type for abstracting the datastore
Motivation
----------
Linq2SQL and EntityFramework have the concept of a "db context" object
which aggregates changes, makes it easier to compose queries and provides
an abstraction of the persistence store.

Modifications
-------------
Added a DbContext class and a concrete example that abstracts the
beer-sample bucket and provides strongly typed properties for the document
types in that bucket.

Results
-------
There is a now a higher level object for interacting with a bucket. It's
possible to inherit from this context object with a concrete
representations. Expect things like change tracking and tooling in future
commits. Also, expect the interface and DbContext to change quite a bit in future
releases especially around construction/destruction.