Support for GROUP BY, HAVING, and aggregates
Motivations
-----------
Allow LINQ group clauses and aggregates to generate valid N1QL statements.
Should support simple aggregates directly against a query, or more
advanced aggregates utilizing groups.  Should also support treating where
clauses after the group clause as HAVING statements.

Modifications
-------------
For simple aggregates, added AggregateFunction to QueryPartsAggregator.
This function wraps SelectParts in the select clause.

Added handling to N1QlQueryVisitor for group clauses.  When grouping is
performed in a subquery as part of the MainFromClause, the group is
flattened onto the main query.  Created KeyExpressionTransformer and
MultiKeyExpressionTransformer classes to transform references to the group
keys after the query has been flattened.  When where clauses are
encountered after flattening, they are added to HavingParts instead of
WhereParts on the QueryPartsAggregator.

Introduced a new N1QlQueryType, Aggregate, which returns a simple
statement such as COUNT(*) from the QueryPartsAggregator.  This is used
when ReLinq generates a subquery in the select clause to perform one part
of a grouped aggregate.

For testing, simple aggregates always return immediately and don't provide
a method to capture the query.  Added the BucketQueryExecutorEmulator
class and the CreateQueryable function to support capturing query text
after the call is complete.  Created tests for all new functionality.

Results
-------
LINQ can now be used to generate queries with GROUP BY and HAVING clauses,
as well as the standard COUNT, COUNT distinct, SUM, AVG, MIN, and MAX
aggregate functions.