Support for date/time constants and comparisons
Motivation
----------
Working with date/time values in documents is a very common use case.  For
example, querying documents modified since a certain date.  Or invoices
within a certain date range.  To support this via LINQ, support for
including date/time constants and for comparing these constants to
document properties is needed.

Modifications
-------------
Before visiting expressions with the N1QL expression tree visitor, they
are first passed through a TransformingExpressionTreeVisitor registered
with a DateTimeComparisonExpressionTransformer.  This converts any
DateTime values on binary comparison expressions to method calls that
convert them to UnixMilliseconds objects.

UnixMilliseconds structures act in .Net as a simple wrapper for a
DateTime, since that is still the most efficient way to work with them in
.Net.  However, the UnixMillisecondsMethodCallTranslator can now detect
this conversion and treat it as a call to STR_TO_MILLIS in N1QL.

The UnixMilliseconds structure is declared internal, and isn't designed
for public use.  It's only intended for use inside expression trees.
However, it does have some basic methods in case they are needed for unit
testing.

For constants, the configured ITypeSerializer for Couchbase is used to
serialize the constant to a string (normally ISO8601) before writing to
the N1QL query.

Results
-------
Date/time constants may now be freely placed in N1QL queries as ISO8601
strings via LINQ.  These constants and/or document properties can now be
used on either side of <, >, <=, >=, =, and <> operators.

Notes
-----
For this basic implementation where only comparisons are supported, using
a TransformingExpressionTreeVisitor may be a bit overkill.  However, it
should provide a more powerful platform for supporting more advanced
functionality in the future, such as date arithmetic.