LINQ-12: "Narrow" the public API - make some public Types internal
Motivation
----------
The exposed public API of the Linq2Couchbase provider is the query API and
the DbContext; this makes other types that are not intended to be used in
a standalone internal.

Modifications
-------------
BucketQueryExecutor, BucketQueryable, QueryFactory, and QueryParserHelper
are no longer public Types.

Results
-------
It's now easier to make non-backward compatible changes as well as coax
users of the API torwards the intended public API.