LINQ-9: Remove ClusterHelper reference from N1QLQueryModelVistor
Motivation
----------
The Linq project had a hard reference to the ClusterHelper; this commit
removes the reference and instead allows the ClientConfiguration to be
injected as a reference.

Modification
------------
The CLusterHelper reference was removed and instead ClientConfiguration is
injected from the higher-level objects.

Result
------
ClusterHelper is now no longer a required reference; you do not need to
call ClusterHelper.Initialize() before executing a query.