LINQ-20: Add Save and Remove to IBucketContext
Motivation
----------
Adds an additional WRITE API for users who do not want to do change
tracking (forthecoming).

Modifications
-------------
Added Save and Remove to IBucketContext and provided implementation in
BucketContext. Added a DocId attribute so the the key for the document Add
Save and Remove to IBucketContextcan be defined from one of the fields; I
considered making a string field called "_id" the default key-field, but
removed that logic - it may come back though. I split the
BucketContextTests into integrated and UT test classes. Finally, added a
couple of exceptions for handling specific use cases such as KeyNotFound
and when a key (id) field cannot be found. Note that Save and Remove throw
exceptions; so either it succeeded or it failed and the app needs to
handle the exception and failure case.

Result
------
You can now Save and Remove documents from the BucketContext.