Refactor error handling for Remove and Save
Motivation
----------
If BucketContext.Remove or Save fail throw a CouchbaseWriteException for
consistency. The CouchbaseWriteException is a wrapper around an
IOperationResult implementation and exposes the ResponseStatus and Message
fields.

Modifications
-------------
Add CouchbaseWriteException class, refactor Save and Remove to throw it
when an operation results in a non-success. Also, added PropertyInfo
caching for KeyAttribute property and fixed bug in GetDocumentId which
returned the name of the key field and not the actual value of the key.

Results
-------
If Save or Remove fail, a CouchbaseWriteException will be thrown with the
information from the IOperationResult.