Issue #62: Fix N1Ql namespace conflict with Couchbase SDK
Motivation
----------
The N1Ql class in Linq2Couchbase is conflicting with the Couchbase.N1Ql
namespace from the SDK.  If both Couchbase and Couchbase.Linq namespaces
are imported, the compiler is having trouble differentiating when
referencing N1Ql.

Modifications
-------------
Renamed N1Ql to N1QlFunctions.  This name is consistent with Entity
Framework, which has previously used EntityFunctions and now uses
DbFunctions and SqlFunctions.

Results
-------
The namespace conflict is now fixed.