Issue #68: Fix handling for queries without a new object select projection
Motivation
----------
Queries that don't include a new object in the select projection (i.e.
`select new {...}`) are failing to deserialize.  The result is returned by
N1QL as an array of objects with a $1 property, and the expected value is
stored in this property.  However, LINQ deserialization is expecting an
array of the plain value, without the intermediate wrapper object.

Modifications
-------------
Added a property to N1QlQueryModelVisitor, ResultExtractionRequired, which
indicates to the BucketQueryExecutor that a property extraction is
required during deserialization.  This property is always named "result".
BucketQueryExecutor now uses this property to recognize that it needs to
deserialize an array of SimpleResult<T> instead of any array of T, then
extracts the result property.

Repurposed "ArrayPropertyExtractionPart", which was formerly used to
perform a similar function when dealing with array subqueries and Any/All.
It is now "PropertyExtractionPart", and is used for array subqueries as
well as main queries where property extraction is required.

Previously, aggregates were manually having property extraction performed
on them.  They now use the new property extraction approach, which
simplifies the system and should make things less confusing going forward.