Issue #76: Fixed string concatenation for non-strings and ToString
Motivation
----------
When building compound keys for joins and nests, a common use case may be
to combine strings with non-string properties such as numbers and GUIDs.

Modifications
-------------
Changed handling of binary + expression to use string concatenation
operator if either operand is a string.  Previously it was only doing this
if both operands were strings.

Added a method call translator for .ToString() that uses the N1QL
TOSTRING() function.  This required making a tweak to the
DefaultMethodCallTranslatorProvider to check for method call translators
against virtual base methods.  This allowed registering the method call
translator only on Object.ToString while applying to all inherited
types.

Results
-------
String handling for non-string types now behaves in the way a user would
expect.