Update to Couchbase SDK 2.2.2
Motivation
----------
Features are required from SDK 2.2.2, but when they were developed 2.2.2
wasn't released yet.  Therefore, we still have relative path references to
the SDK, instead of using the NuGet package.

Modifications
-------------
Updated NuGet references to SDK 2.2.2, and modified the assembly
references to use the NuGet package instead of a relative path to the SDK
repository.

Results
-------
The Linq2Couchbase repository can now be cloned and built without also
cloning the SDK.  This also ensures that testing is done with the correct
NuGet version of the SDK.