Issue #89: Fixed prefiltered array ALL subqueries
Motivation
----------
Query standards for extent names changed slightly for Couchbase Server
4.1.  We can no longer reuse extent names outside of an ARRAY statement
that were defined inside the ARRAY statement.  This was occuring when a
nested array was being filtered with a Where clause and then had an All
result operator applied to it.

Modifications
-------------
Added new feature to the ExtentNameProvider, GenerateNewExtentName.  This
is used to create a new name for an existing extent.  This allows us to
reference the original IQuerySource in the All predicate, but do so using
a different extent name.

We then store this extent name in
QueryPartsAggretator.PropertyExtractionPart so that it can be used
building the EVERY clause.

Results
-------
Filtered nested arrays can now have a All result operator applied to them
when running against Couchbase Server 4.1.