Issue #88: Support for string enumeration comparisons
Motivation
----------
A common use case for enumerations is to store them in the JSON document
as strings rather than integers.  With Json.Net, this is usually done with
StringEnumConverter.  In order to filter N1QL documents based on these
enumerations in WHERE clauses, the outputed N1QL query must compare using
the string representation instead of the integer representation.

Modifications
-------------
Whenever a == or != comparison is performed against an enumeration
constant, convert the expression in EnumComparisonExpressionTransformer
to use the enumeration type instead of the base integer type.  This allows
the constant to be recognized as an enumeration constant later by
N1QlExpressionTreeVisitor.

When an enumeration constant is encountered by N1QlExpressionTreeVisitor,
use the configured ITypeSerializer to determine how to output the value to
N1Ql.  This is done by first serializing to JSON, then deserializing back
to a .Net type.  This gets a string or integer value, depending on how the
type serializer is configured.  This can then be output to the N1QL query
like any other constant.

Results
-------
Equal and not equal comparisons of enumeration values are now supported
when using a StringEnumConverter.  This should theoretically extend to any
other non-standard JSON conversion, so long as the non-standard
serialization is applied directly at the enumeration type.  It would not
work, for example, if a custom serializer was applied to the document
itself that changed how the enumeration property is serialized.

Note that greater than and less than comparisons are not supported with
StringEnumConverter.  They are still supported when working with numeric
serialization.  However, they simply don't translate to N1QL from .Net as
a string, because the .Net comparison greater than or less than comparison
would still be based on the numeric value.  In N1QL, it would be based on
the string value from the serialization layer.