Issue #107: Support for subquerys against constant lists
Motivation
----------
A desirable use case is to use the .Contains function against a static
list to test if a document attribute matches at least one item in the
list.  This can be implemented with the N1QL IN keyword.

Modifications
-------------
Relinq treats this is a subquery, with a constant value implementing
IEnumerable as the expression on the from clause.  Detect this case, and
render a static array to N1QL when it is encountered.  We can treat this
like any other array-type subquery, which includes support for the
.Contains method.

Results
-------
array.Contains(x) can now be used within a query, where array is constant
in relation to the query.  Note that array can be dynamically built before
the query, it is simply constant within the query's execution plan and
does not vary based on data returned within the query.

Also, due to the implementation approach, it also supports more advanced
subqueries against constant arrays.  For example, it is possible to filter
or perform select projections against the constant array before the
.Contains statement.  In the future, it might be worthwhile to optimize
these cases by performing filtering and projections before writing the
array to the N1QL query.  However, it is functional now.