Issue #102: Aggregates after select projections
Motivation
----------
If you apply a select clause in LINQ, and then follow this with an
aggregate such as .Count(), the generated N1QL is invalid.  It is
including syntax like "`a` as `a`, `b` as `b`" inside the COUNT()
function.

Modifications
-------------
Recognize when a new object select statement is used within an aggregate.
When encountered, use the JSON {...} new object syntax in N1QL.

Results
-------
N1QL created is now valid and runs successfully.  Note that using the JSON
new object syntax like COUNT({a: b}) is more valid than simply always
using COUNT(*).  This is because it will work with other operations like
COUNT(DISTINCT {a: b}).