Don't serializer __interceptors field on proxies
Motivation
----------
When document proxy generation is enabled, the proxied documents gain an
extra __interceptors attribute when they are serialized back to JSON.

Modifications
-------------
Created DocumentProxyGenerator which overrides the behavior of the default
ClassProxyGenerator so that IgnoreDataMemberAttribute is applied to the
emitted __interceptors field.

Created DocumentProxyBuilder which overrides DefaultProxyBuilder to use
DocumentProxyGenerator instead of ClassProxyGenerator, and inject it into
the default DocumentProxyManager.

Created a unit test that ensures that "__interceptors" is not found in the
output JSON for a test proxy document.

Results
-------
__interceptors attribute is no longer serialized when document proxies are
written back to Couchbase.